cos-logger as a separate module.
