+++ tekton-catalog/cos-logger/go.mod
@@ -1,0 +1,5 @@
+module github.com/kubeflow/kfp-tekton/tekton-catalog/cos-logger
+
+require github.com/IBM/ibm-cos-sdk-go v1.8.0
+
+go 1.13
