+++ tekton-catalog/pipeline-loops/pkg/reconciler/pipelinelooprun/controller.go
@@ -21,12 +21,14 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
+	"strconv"
 	"syscall"
 
 	"github.com/kubeflow/kfp-tekton/tekton-catalog/pipeline-loops/pkg/apis/pipelineloop"
 	pipelineloopv1alpha1 "github.com/kubeflow/kfp-tekton/tekton-catalog/pipeline-loops/pkg/apis/pipelineloop/v1alpha1"
 	pipelineloopclient "github.com/kubeflow/kfp-tekton/tekton-catalog/pipeline-loops/pkg/client/injection/client"
 	pipelineloopinformer "github.com/kubeflow/kfp-tekton/tekton-catalog/pipeline-loops/pkg/client/injection/informers/pipelineloop/v1alpha1/pipelineloop"
+	cl "github.com/scrapcodes/cos-logger"
 	pipelineclient "github.com/tektoncd/pipeline/pkg/client/injection/client"
 	runinformer "github.com/tektoncd/pipeline/pkg/client/injection/informers/pipeline/v1alpha1/run"
 	pipelineruninformer "github.com/tektoncd/pipeline/pkg/client/injection/informers/pipeline/v1beta1/pipelinerun"
@@ -34,12 +36,35 @@ import (
 	pipelinecontroller "github.com/tektoncd/pipeline/pkg/controller"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
+	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
 	kubeclient "knative.dev/pkg/client/injection/kube/client"
 	"knative.dev/pkg/configmap"
 	"knative.dev/pkg/controller"
 	"knative.dev/pkg/logging"
+	"knative.dev/pkg/system"
 )
+
+func load(ctx context.Context, kubeClientSet kubernetes.Interface, o *cl.ObjectStoreLogConfig) error {
+	configMap, err := kubeClientSet.CoreV1().ConfigMaps(system.Namespace()).
+		Get(ctx, "object-store-config", metaV1.GetOptions{})
+	if err != nil {
+		return err
+	}
+	if o.Enable, err = strconv.ParseBool(configMap.Data["enable"]); err != nil || !o.Enable {
+		return err
+	}
+
+	o.AccessKey = configMap.Data["accessKey"]
+	o.SecretKey = configMap.Data["secretKey"]
+	o.Region = configMap.Data["region"]
+	o.ServiceEndpoint = configMap.Data["serviceEndpoint"]
+	o.DefaultBucketName = configMap.Data["defaultBucketName"]
+	o.CreateBucket = false
+	o.Token = configMap.Data["token"]
+	return nil
+}
 
 // NewController instantiates a new controller.Impl from knative.dev/pkg/controller
 func NewController(namespace string) func(context.Context, configmap.Watcher) *controller.Impl {
@@ -60,10 +85,17 @@ func NewController(namespace string) func(context.Context, configmap.Watcher) *c
 			pipelineLoopLister:    pipelineLoopInformer.Lister(),
 			pipelineRunLister:     pipelineRunInformer.Lister(),
 		}
-		objectStoreLogger := Logger{
+		loggerConfig := cl.ObjectStoreLogConfig{}
+		objectStoreLogger := cl.Logger{
 			MaxSize: 1024 * 100, // TODO make it configurable via a configmap.
 		}
-		err := objectStoreLogger.LoadDefaults(ctx, kubeclientset)
+		err := load(ctx, kubeclientset, &loggerConfig)
+		if err == nil {
+			err = objectStoreLogger.LoadDefaults(loggerConfig)
+			if err == nil {
+				_ = objectStoreLogger.LogConfig.CreateNewBucket(loggerConfig.DefaultBucketName)
+			}
+		}
 		if err == nil && objectStoreLogger.LogConfig.Enable {
 			logger.Info("Loading object store logger...")
 			w := zapcore.NewMultiWriteSyncer(
